feat(parse): move signed event parsing to file
