+++ src/event_message/mod.rs
@@ -119,84 +119,7 @@ impl Verifiable for SignedEventMessage {
 
 const JSON_SIG_DELIMITER: &str = "\n";
 
-pub mod parse_signed_event {
-    use super::{AttachedSignaturePrefix, EventMessage, SignedEventMessage};
-    use crate::prefix::attached_signature::{b64_to_num, parse_attached_signature::signature};
-    use nom::{branch::*, combinator::*, error::ErrorKind, multi::*, sequence::*};
-
-    fn json_message(s: &str) -> nom::IResult<&str, EventMessage> {
-        let mut stream =
-            serde_json::Deserializer::from_slice(s.as_bytes()).into_iter::<EventMessage>();
-        match stream.next() {
-            Some(Ok(event)) => Ok((&s[stream.byte_offset()..], event)),
-            _ => Err(nom::Err::Failure((s, ErrorKind::IsNot))),
-        }
-    }
-
-    fn cbor_message(s: &str) -> nom::IResult<&str, EventMessage> {
-        let mut stream =
-            serde_cbor::Deserializer::from_slice(s.as_bytes()).into_iter::<EventMessage>();
-        match stream.next() {
-            Some(Ok(event)) => Ok((&s[stream.byte_offset()..], event)),
-            _ => Err(nom::Err::Failure((s, ErrorKind::IsNot))),
-        }
-    }
-
-    fn message(s: &str) -> nom::IResult<&str, EventMessage> {
-        alt((json_message, cbor_message))(s)
-    }
-
-    /// extracts the count from the sig count code
-    fn sig_count(s: &str) -> nom::IResult<&str, u16> {
-        let (rest, t) = tuple((
-            map_parser(
-                nom::bytes::complete::take(2u8),
-                tuple((
-                    nom::bytes::complete::tag("-"),
-                    nom::bytes::complete::tag("A"),
-                )),
-            ),
-            map(nom::bytes::complete::take(2u8), |b64_count| {
-                b64_to_num(b64_count).map_err(|_| nom::Err::Failure((s, ErrorKind::IsNot)))
-            }),
-        ))(s)?;
-
-        Ok((rest, t.1?))
-    }
-
-    /// called on an attached signature stream starting with a sig count
-    fn signatures(s: &str) -> nom::IResult<&str, Vec<AttachedSignaturePrefix>> {
-        let (rest, t) = tuple((sig_count, many0(signature)))(s)?;
-        if t.0 as usize != t.1.len() {
-            Err(nom::Err::Failure((s, ErrorKind::Count)))
-        } else {
-            Ok((rest, t.1))
-        }
-    }
-
-    fn signed_message(s: &str) -> nom::IResult<&str, SignedEventMessage> {
-        let (rest, t) = nom::sequence::tuple((message, signatures))(s)?;
-        Ok((rest, SignedEventMessage::new(&t.0, t.1)))
-    }
-
-    #[test]
-    fn test() {
-        use crate::prefix::SelfSigningPrefix;
-        assert_eq!(sig_count("-AAA"), Ok(("", 0u16)));
-        assert_eq!(
-            sig_count("-AABextra data and stuff"),
-            Ok(("extra data and stuff", 1u16))
-        );
-
-        assert_eq!(
-            signatures("-AABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
-            Ok(("", vec![AttachedSignaturePrefix {
-                index: 0,
-                sig: SelfSigningPrefix::Ed25519Sha512([0u8; 64].to_vec())
-            }]))
-        )
-    }
-}
+pub mod parse;
 
 pub fn parse_signed_message_json(message: &str) -> Result<SignedEventMessage, Error> {
     let parts: Vec<&str> = message.split(JSON_SIG_DELIMITER).collect();
